Use ideal state to compute segment assignment
Before this patch, we used the external view to compute segment
assignment. If a server if offline, it does not appear in external
view. As such, an offline server will be considered empty when
assigning segments and will be assigned all new segments until it
comes back online.

This patch fixes this behavior by using the ideal state instead of the
external view for segment assignment.